requirements.txt
asgiref==3.8.1
bleach==6.2.0
crispy-bootstrap5==2024.10
dj-database-url==0.5.0
Django==5.1.7
django-allauth==65.4.1
django-crispy-forms==2.3
django-summernote==0.8.20.0
gunicorn==20.1.0
psycopg2==2.9.10
setuptools==76.0.0
sqlparse==0.5.3
webencodings==0.5.1
whitenoise==6.5.0
